+++ heptools/container/partial_set.py
@@ -1,3 +1,4 @@
+# TODO remove
 from __future__ import annotations
 
 from typing import Iterable
@@ -8,23 +9,22 @@ import numpy as np
 from ..aktools import AnyArray
 
 
-# TODO numba
 class PartialSet:
     sort_kind = 'mergesort'
 
     def _zip(self, *indices: AnyArray):
         return ak.zip(indices) if len(indices) > 1 else indices[0]
 
-    def __init__(self, value: bool | Iterable[bool], *indices: AnyArray, default = False):
+    def __init__(self, value: bool | Iterable[bool], *indices: AnyArray, default=False):
         indices = np.array(self._zip(*indices))
-        self._in = np.empty(0, dtype = indices.dtype)
+        self._in = np.empty(0, dtype=indices.dtype)
         self._out = default
         if isinstance(value, bool):
             if value != default:
                 self._in = indices
         else:
             self._in = indices[value != default]
-        self._in.sort(kind = self.sort_kind)
+        self._in.sort(kind=self.sort_kind)
 
     @classmethod
     def new(cls, _in: np.ndarray, _out: bool, sort: bool = False):
@@ -32,13 +32,13 @@ class PartialSet:
         new._in = _in
         new._out = _out
         if sort:
-            new._in.sort(kind = cls.sort_kind)
+            new._in.sort(kind=cls.sort_kind)
         return new
 
     def __invert__(self):
         return self.new(self._in, not self._out)
 
-    def __and__(self, other: PartialSet) -> PartialSet: # TODO check type of self._in and other._in
+    def __and__(self, other: PartialSet) -> PartialSet:
         if isinstance(other, PartialSet):
             if self._out:
                 if other._out:
@@ -60,23 +60,24 @@ class PartialSet:
 
     def __xor__(self, other: PartialSet) -> PartialSet:
         if isinstance(other, PartialSet):
-            return self.new(np.setxor1d(self._in, other._in, assume_unique = True), self._out is not other._out, True)
+            return self.new(np.setxor1d(self._in, other._in, assume_unique=True), self._out is not other._out, True)
         return NotImplemented
 
     def __add__(self, other: PartialSet) -> PartialSet:
         if isinstance(other, PartialSet):
             if not self._out == other._out:
-                raise ValueError('cannot add two partial sets with different default values')
+                raise ValueError(
+                    'cannot add two partial sets with different default values')
             return self.new(np.concatenate((self._in, other._in)), self._out or other._out, True)
         return NotImplemented
 
     def __call__(self, *indices: AnyArray):
         indices = np.array(self._zip(*indices))
-        return np.isin(indices, self._in, invert = self._out)
+        return np.isin(indices, self._in, invert=self._out)
 
     def __len__(self):
         return len(self._in)
 
     def unique(self):
-        self._in, count = np.unique(self._in, return_counts = True)
+        self._in, count = np.unique(self._in, return_counts=True)
         return self._in[count > 1]
